Move testing helpers to the test
