--- albums/helpers_test.go
@@ -1,12 +1,0 @@
-package albums_test
-
-import "testing"
-
-func assertExpectedError(isErrExpected bool, err error, t *testing.T) {
-	if isErrExpected && err == nil {
-		t.Fatalf("error was expected, but not produced")
-	}
-	if !isErrExpected && err != nil {
-		t.Fatalf("error was not expected, err: %s", err)
-	}
-}
